Moved the code to look for old graphs into a GraphUpdater.ts
